requirements.txt
Django
Pillow
psycopg2-binary
djangorestframework
drf-yasg
django-phonenumber-field
phonenumbers